requirements.txt
fastapi>=0.78.0
pydantic>=1.9.1
uvicorn>=0.18.2
tortoise-orm>=0.19.1
tortoise-orm[asyncpg]>=0.19.1
celery>=5.2.7
redis>=4.3.4
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5